don't process files that were touched in the last 60 mins

+++ ftpSummaryStatsScript/depo_ftp_to_staging.sh
@@ -14,7 +14,8 @@ harmonise_dir=$3
 
 # need to chmod so everyone can read
 for f in $source_dir/*; do
-        if [ -e "$f" ]; then
+        # if the file exists and has not been modified in the last 60 mins
+        if [ -e "$f" ] && [ `find "$f" -mmin +60` ]; then
                 rsync -rv --chmod=Du=rwx,Dg=rx,Do=rx,Fu=rw,Fg=r,Fo=r $f $staging_dir/
                 mv -v $f $harmonise_dir/
         fi
